GraphQL endpoint at providers level
